fix search bar issues
